Fixing PMD issues
