+++ kernel-core/src/main/resources/i18n/io/zephyr/kernel/launch/KernelOptions_en_US.properties
@@ -1,3 +1,12 @@
 options.file.located=Found file ''{0}''
 options.file.locating=Searching for {0} in realm {1}...
 options.homedirectory.exists=Sunshower Home directory ''{0}'' already exists...skipping
+
+filesystem.resolve.root.system_properties.begin=Attempting to resolve file system root from system properties, key "zephyr.options.home"
+filesystem.resolve.root.system_properties.success=Successfully resolved file system root from system properties, key "zephyr.options.home", value ''{0}''
+filesystem.resolve.root.system_properties.doesnt_exist=System property ''{0}'' is not set.  Checking for ZEPHYR_HOME in environment variables
+filesystem.resolve.root.does_not_exist=Located a value ''{0}'' from {1}, but no file at that location exists
+filesystem.resolve.root.is_not_directory=File ''{0}'' located in {1} exists, but it is not a directory
+filesystem.resolve.root.permissions_failed=File ''{0}'' located in {1} exists, but the current user ''{2}'' does not have {3} permissions
+filesystem.resolve.root.env_var.doesnt_exist=ENV VARIABLE ''{0}'' is not set.  Attempting resolution from options
+filesystem.resolve.root.env_var.success=Successfully resolved file system root from environment, key "ZEPHYR_HOME", value ''{0}''
